# other/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API Server of Codern",
        "title": "Codern API Server",
        "contact": {},
        "version": "0.0.0"
    },
    "paths": {
        "/auth/google": {
            "get": {
                "description": "Get an url to signin with the Google account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get Google auth URL",
                "responses": {}
            }
        },
        "/auth/google/callback": {
            "get": {
                "description": "A callback route for Google OAuth to redirect to after signing in",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in with Google",
                "responses": {}
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get an authenticated user data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get an user data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/auth/signin": {
            "post": {
                "description": "Sign in with email \u0026 password provided by the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in with self provider",
                "responses": {}
            }
        },
        "/auth/signout": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuths": []
                    }
                ],
                "description": "Sign out and remove a sid cookie header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign out",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/file/user/{userId}/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get an user profile image from internal file system by proxy",
                "produces": [
                    "image/png",
                    "image/jpeg",
                    "image/gif"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Get an user profile image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/file/workspaces/{workspaceId}/assignments/{assignmentId}/{subPath}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a workspace detail markdown from internal file system by proxy",
                "produces": [
                    "image/png",
                    "image/jpeg",
                    "image/gif"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Get a workspace detail markdown",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Workspace ID",
                        "name": "workspaceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Assignment ID",
                        "name": "assignmentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sub path",
                        "name": "subPath",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/file/workspaces/{workspaceId}/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a workspace profile image from internal file system by proxy",
                "produces": [
                    "image/png",
                    "image/jpeg",
                    "image/gif"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Get a workspace profile image",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Workspace ID",
                        "name": "workspaceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/workspaces": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all workspaces",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "List workspaces",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "participants"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Specific fields to include in the response",
                        "name": "fields",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/workspaces/{workspaceId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a workspace from workspace id on path parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Get a workspace",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workspace ID",
                        "name": "workspaceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "participants"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Specific fields to include in the response",
                        "name": "fields",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/workspaces/{workspaceId}/assignments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all assignment from a workspace id on path parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "List assignment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workspace ID",
                        "name": "workspaceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/workspaces/{workspaceId}/assignments/{assignmentId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get an assignment from workspace id on path parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Get an assignment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workspace ID",
                        "name": "workspaceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Assignment ID",
                        "name": "assignmentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/workspaces/{workspaceId}/assignments/{assignmentId}/submissions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all submission from a workspace id on path parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "List submission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workspace ID",
                        "name": "workspaceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Assignment ID",
                        "name": "assignmentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Submit a submission of the assignment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Create a new submission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workspace ID",
                        "name": "workspaceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Assignment ID",
                        "name": "assignmentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.CreateSubmissionPayload"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sid",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "payload.CreateSubmissionPayload": {
            "type": "object",
            "required": [
                "language",
                "sourceCode"
            ],
            "properties": {
                "language": {
                    "type": "string"
                },
                "sourceCode": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}